changed folder structure of course and enrollment
