+++ backend/features/enrollment/enrollment.validation.js
@@ -1,0 +1,6 @@
+// validators/enrollmentValidator.js
+const { body } = require('express-validator');
+
+exports.enrollValidator = [
+  body('courseId').notEmpty().withMessage('Course ID is required')
+];
